animation frame
not working, but some groundwork laid.  changed most objects over to be
positioned in reference to the sprite sheet, and other objects to be
referenced from those in an attempt to eliminate magic numbers. added
minion animated object to initialize, draw, and update, but the coords
are still messed up so the image does not show properly.